presentation-api/controlling-ua/PresentationAvailability_onchange.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Monitoring the list of available presentation displays.</title>
<link rel="author" title="Tomoyuki Shimizu" href="https://github.com/tomoyukilabs">
<link rel="help" href="http://w3c.github.io/presentation-api/#monitoring-the-list-of-available-presentation-displays">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="common.js"></script>

<p id="notice">The test passes if a "PASS" result appears.<br></p>

<script>
    // prevent the default timeout
    setup({explicit_timeout: true});

    var notice = document.getElementById('notice');

    promise_test(function(t) {
      // clean up the instruction notice when the test ends
      t.add_cleanup(function() {
          notice.style.display = 'none';
      });

      // initialize a presentation request
      var request = new PresentationRequest(presentationUrls);

      var availability, previousState, eventWatcher;

      // set timeout to observe the presentation availability
      t.step_timeout(function() {
        t.force_timeout();
        t.done();
      }, 90000);

      var wait = t.step_func(function(a) {
        availability = a;

        // save the current value of the presentation availability
        previousState = availability.value;

        // add an instruction notice
        notice.innerHTML += 'Please make your presentation display '
                          + (previousState ? 'unavailable' : 'available')
                          + ' and wait for a minute...';

        // wait until a "change" event is fired
        if(!eventWatcher)
          eventWatcher = new EventWatcher(t, availability, ['change', 'change']);
        return eventWatcher.wait_for('change');
      });

      var check = t.step_func(function(evt) {
        notice.innerHTML += 'done.<br>';

        // check the event and its attributes
        assert_true(evt.isTrusted && !evt.bubbles && !evt.cancelable && evt instanceof Event, 'A simple event is fired.');
        assert_equals(evt.type, 'change', 'The event name is "change".');
        assert_equals(evt.target, availability, 'event.target is the presentation availability.');
        assert_not_equals(previousState, availability.value, 'Value of the presentation availability is changed.');
        return evt.target;
      });

      // check the change of PresentationAvailability.value twice; "true to false" and "false to true"
      return request.getAvailability().then(wait).then(check).then(wait).then(check);
    });
</script>
